Use `dirname` instead of trimming on `/`
When running on Windows, buildkite can provide backslash-separated paths
for the script name, causing the variable substitution here to fail to
trim off the basename of the invoked hook file.  This changes the hooks
to instead use `dirname`, which on these windows systems works with both
forward and backward slash separated names.